refactor-integrations (#112)
